% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Enrichment-class}
\alias{Enrichment-class}
\title{Enrichment-class}
\description{
An S4 class to store enrichment analysis results, including
the enrichment results of input and background protein lists.
}
\section{Slots}{

\describe{
\item{\code{input}}{An object containing enrichment results for the input set}

\item{\code{background}}{An object containing enrichment results for the background set}
}}

\examples{
# Create an Enrichment object
obj <- methods::new("Enrichment",
  input      = data.frame(Gene = c("TP53", "BRCA1")),
  background = data.frame(Gene = c("TP53", "BRCA1", "EGFR", "MYC"))
)
show(obj)
}
