% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{getBackground}
\alias{getBackground}
\alias{getBackground,Enrichment-method}
\title{Get background}
\usage{
getBackground(object)

\S4method{getBackground}{Enrichment}(object)
}
\arguments{
\item{object}{An Enrichment object}
}
\value{
The background data frame
}
\description{
Accessor for the background slot of an Enrichment object
}
\examples{
obj <- methods::new("Enrichment",
  input      = data.frame(Gene = c("TP53", "BRCA1")),
  background = data.frame(Gene = c("TP53", "BRCA1", "EGFR"))
)
getBackground(obj)
}
