% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{getInput}
\alias{getInput}
\alias{getInput,Enrichment-method}
\title{Get input}
\usage{
getInput(object)

\S4method{getInput}{Enrichment}(object)
}
\arguments{
\item{object}{An Enrichment object}
}
\value{
The input data frame
}
\description{
Accessor for the input slot of an Enrichment object
}
\examples{
obj <- methods::new("Enrichment",
  input      = data.frame(Gene = c("TP53", "BRCA1")),
  background = data.frame(Gene = c("TP53", "BRCA1", "EGFR"))
)
getInput(obj)
}
