% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tissue_tpm_gtex}
\alias{tissue_tpm_gtex}
\title{GTEx tissue-average TPM matrix}
\format{
A numeric matrix with genes (Ensembl gene IDs) as rows and tissues as columns.
}
\source{
The Genotype-Tissue Expression (GTEx) Project
\url{https://www.gtexportal.org/}
GTEx Consortium (2013). \emph{The Genotype-Tissue Expression (GTEx) project}.
\emph{Nature Genetics}, 45(6). \doi{10.1038/ng.2653}
}
\usage{
tissue_tpm_gtex
}
\description{
A numeric matrix containing tissue-averaged transcript-per-million (TPM)
expression values from the GTEx project. Rows correspond to genes
(Ensembl gene IDs) and columns correspond to human tissues.
}
\keyword{datasets}
