% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ee_tb}
\alias{ee_tb}
\title{Human Protein Atlas tissue expression database}
\format{
A data frame with \code{n} rows and the following columns:
\describe{
\item{Gene}{HGNC gene symbol}
\item{Ensembl}{Ensembl gene identifier}
\item{EntrezID}{NCBI Entrez ID}
\item{Uniprot}{UNIPROT ID}
\item{Is_secreted}{Whether the protein is a predicted secretory protein  or not}
\item{Tissue.specificity}{expression category, one of \code{"Tissue enhanced"},
\code{"Group enriched"}, or \code{"Tissue enriched"}}

\item{tissue}{tissue names}
}
}
\source{
Human Protein Atlas \url{https://www.proteinatlas.org}.
Uhlen M et al. (2017). A pathology atlas of the human cancer
transcriptome. \emph{Science}, 357(6352).
\doi{10.1126/science.aan2507}
}
\usage{
ee_tb
}
\description{
A data frame containing tissue-specific gene expression data
derived from the Human Protein Atlas (HPA).
}
\keyword{datasets}
