% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tissue_map}
\alias{tissue_map}
\title{An example of the data frame the for the 'tissue_groups' parameter of
\code{geneClassification}.}
\format{
An example of the data frame the for the 'tissue_groups' parameter of
\code{geneClassification}. This data frame has two columns named
\code{Tissue} and \code{Group}, defining predefined tissue groupings for
group-enriched analysis.
}
\usage{
tissue_map
}
\description{
An example of the data frame the for the 'tissue_groups' parameter of
\code{geneClassification}.
}
\keyword{datasets}
